seperate BulkDialog for bulk delete
